+++ CHANGELOG.md
@@ -5,6 +5,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.0] - 2026-03-22
+
+## Added
+ - function Get-BrevoNote
+ - function New-BrevoNote
+ - function Remove-BrevoNote
+ - function Update-BrevoNote
+
+## [0.4.1] - 2026-03-22
+
 ### Added
 
 - function Update-BrevoContactAttribute